% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afr_va.R
\name{afr_va}
\alias{afr_va}
\title{Function minimized in model}
\usage{
afr_va(x, A, nassay, nreceptor, pathway, penalty_method, alpha)
}
\arguments{
\item{x}{x}

\item{A}{A matrix}
}
\value{
eret
}
\description{
\code{afr_va} is the function passed to \code{optim} in \code{er_model}.
}

