% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ar_L5_invitrodb}
\alias{ar_L5_invitrodb}
\title{AR ToxCast Data}
\format{A data.table with 42734 rows and 9 columns}
\usage{
ar_L5_invitrodb
}
\description{
A dataset containing ToxCast model parameters, hit calls, assays, and
chemical info to generate AR Model AUC values. Data is pulled from the
internal release (invitrodb), and subset down to
the minimum needed to run the model.
}
\keyword{datasets}

