% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ar_L5_invitrodb_viability}
\alias{ar_L5_invitrodb_viability}
\title{AR Viability ToxCast Data}
\format{A data.table with 16179 rows and 9 columns}
\usage{
ar_L5_invitrodb_viability
}
\description{
A dataset containing ToxCast model parameters, hit calls, assays, and
chemical info to generate ER Model AUC values. Data is pulled from the
internal release (invitrodb), and subset down to
the minimum needed to run the model. Viability data for TOX21 Antagonist
assays used to filter activity calls.
}
\keyword{datasets}

