% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCcalc.R
\name{AUCcalc}
\alias{AUCcalc}
\title{Calculate the area under the curve}
\usage{
AUCcalc(resmat, pathway, aucscale2 = NULL)
}
\arguments{
\item{resmat}{matrix[nconc rows x nreceptor columns]}
}
\value{
auc numeric vector length nreceptor of AUC values
}
\description{
\code{AUCcalc} calculates the AUC values.
}

