% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{er_L5_prod_ext_v2}
\alias{er_L5_prod_ext_v2}
\title{ER ToxCast Data}
\format{A data.table with 60330 rows and 9 columns}
\usage{
er_L5_prod_ext_v2
}
\description{
A dataset containing ToxCast model parameters, hit calls, assays, and
chemical info to generate ER Model AUC values. Data is pulled from the
October 2015 public release (prod_external_invitrodb_v2), and subset down to
the minimum needed to run the model.
}
\keyword{datasets}

