% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_model.R
\name{er_model}
\alias{er_model}
\title{Calculate the least squares solution for one chemical}
\usage{
er_model(adata, do.debug = F, xmin.plot = 0.01)
}
\arguments{
\item{adata}{cr.mat from \code{prepCR}}

\item{do.debug}{logical to run in debug mode. Default F}

\item{xmin.plot}{numeric. Default 0.01}
}
\value{
resmat matrix[nconc rows x NRECEPTOR columns]
}
\description{
\code{er_model} calculates least squares solution
}

