% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_model_light.R
\name{er_model_light}
\alias{er_model_light}
\title{ER Model Light}
\usage{
er_model_light(dat, chem)
}
\arguments{
\item{dat}{data.table returned by \code{tcpl_to_model_dat}}

\item{chem}{the value of column code to specify the row to perform the calculation on.}
}
\value{
dat_auc data.table of AUC values
}
\description{
\code{er_model_light} combines the returns AUC values for a tcpl input.
}

