% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_model_light_grouped.R
\name{er_model_light_grouped}
\alias{er_model_light_grouped}
\title{ER Model Light Grouped}
\usage{
er_model_light_grouped(dat, group, value)
}
\arguments{
\item{dat}{data.table returned by \code{tcpl_to_model_dat}}

\item{group}{column name such that dat[group == value] will result in a data
table with column code having multiple rows.}

\item{value}{value to subset the group column by}
}
\value{
dat_auc data.table of AUC values
}
\description{
\code{er_model_light_grouped} wraps the function \code{er_model_light} in
lapply to calculate many rows subset by the value of parameter value of
column group.
}
\details{
This is implemented to give bigger chunks to each core in mclapply
  so that the scheduler is not overburdened.
}

