% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptor_score.R
\name{receptor_score}
\alias{receptor_score}
\title{Receptor Score}
\usage{
receptor_score(x)
}
\arguments{
\item{x}{x}
}
\value{
score
}
\description{
\code{receptor_score} calculates the receptor score.
}

