% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcpl_to_model_dat.R
\name{tcpl_to_model_dat}
\alias{tcpl_to_model_dat}
\title{Cast tcpl output to model format}
\usage{
tcpl_to_model_dat(dat, pathway)
}
\arguments{
\item{dat}{output from tcpl level 5 data}

\item{pathway}{string, specify ER or AR}
}
\value{
dat_cast a data.table cast into the correct format for the model
}
\description{
\code{tcpl_to_model_dat} preps tcpl output for model input.
}
\details{
# This function is to take a pipeline formatted data.table
(long format, one m4id per row)
and convert to the form needed by the ER model
(one chemical per row, multiple columns for assays)
Requires a tcplPrepOtpt(tcplLoadData()) formatted data.table
also requires an assay_order (list of assay) so that NAs can be converted
assumed already subset (one unique row per code)
}

