% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcpl_to_model_dat_boot.R
\name{tcpl_to_model_dat_boot}
\alias{tcpl_to_model_dat_boot}
\title{Cast tcpl output to model format}
\usage{
tcpl_to_model_dat_boot(dat, assay_order)
}
\arguments{
\item{dat}{output from \code{toxbootHitParamCI} data}

\item{assay_order}{order of assays}
}
\value{
dat_cast a data.table cast into the correct format for the model
}
\description{
\code{tcpl_to_model_dat} preps toxboot output for model input.
}
\details{
# This function is to take a toxbootHitParamCI formatted data.table
(long format, one m4id bootstrap replicate per row)
and convert to the form needed by the ER model
(one chemical boostrap replicate per row, multiple columns for assays)
Requires an assay_order (list of assay) so that NAs can be converted
assumed already subset (one unique row per code)
}

