% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/via_filter.R
\name{via_filter}
\alias{via_filter}
\title{Filter hit calls in AR dataset by antagonist viability}
\usage{
via_filter(dat, dat_via)
}
\arguments{
\item{dat}{data.table with AR assay results}

\item{dat_via}{data.table with TOX21_AR_BLA_Antagonist_viability and
TOX21_AR_LUC_MDAKB2_Antagonist2_viability results}
}
\value{
dat_via_filt, which is input data.table dat with added column
  modl_ga_via and hitc for TOX21_AR_BLA_Antagonist_ratio and
  TOX21_AR_LUC_MDAKB2_Antagonist2 assays changed to 0 if viability result is
  more potent.
}
\description{
\code{via_filter} removes the hit call for AR antagonist assays if the
viability assay is a more potent hit.
}

