% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_model.R
\name{er_model}
\alias{er_model}
\title{Calculate the least squares solution for one chemical}
\usage{
er_model(dat, code, pathway = "ER", nassay = NULL, conclist = NULL,
  nreceptor = NULL, aucscale1 = NULL, penalty_method = "THRESHOLD",
  alpha = NULL, do.debug = F)
}
\arguments{
\item{dat}{data.table from tcpl_to_model_dat_boot}

\item{code}{string chemical}

\item{pathway}{string}

\item{nassay}{numeric number of assays}

\item{conclist}{numeric vector}

\item{nreceptor}{numeric}

\item{aucscale1}{numeric}

\item{penalty_method, }{character, options are THRESHOLD (default),}

\item{alpha}{numeric, default NULL will be set to pathway specific defaults}

\item{do.debug}{logical to run in debug mode. Default F}
}
\value{
resmat matrix[nconc rows x nreceptor columns]
}
\description{
\code{er_model} calculates least squares solution
}

