% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_model.R
\name{er_model}
\alias{er_model}
\title{Calculate the least squares solution for one chemical}
\usage{
er_model(dat, code, pathway = "ER", nassay = NULL, conclist = NULL,
  nreceptor = NULL, aucscale1 = NULL, penalty_method = "THRESHOLD",
  alpha = NULL, do.debug = F)
}
\arguments{
\item{do.debug}{logical to run in debug mode. Default F}

\item{adata}{cr.mat from \code{prepCR}}
}
\value{
resmat matrix[nconc rows x nreceptor columns]
}
\description{
\code{er_model} calculates least squares solution
}

