% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty.R
\name{penalty}
\alias{penalty}
\title{Penalty}
\usage{
penalty(x, penalty_method = "THRESHOLD")
}
\arguments{
\item{x}{x}

\item{penalty_method}{character specifying which algorithm to use when
calculating the penalty. Default is "THRESHOLD" with other options "RIDGE"
and "LASSO".}
}
\value{
value
}
\description{
\code{penalty} returns value of penalty.
}

