% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepCR.R
\name{prepCR}
\alias{prepCR}
\title{\code{prepCR} preps the concentration-response matrix for a single chemical}
\usage{
prepCR(dat, chem, conclist, nassay, modl_ga_cols, modl_tp_cols, modl_gw_cols)
}
\arguments{
\item{dat}{data.table}

\item{chem}{value of \code{code} column to specify the row}

\item{conclist}{numeric vector}

\item{nassay}{integer length 1}

\item{modl_ga_cols}{character vector}

\item{modl_tp_cols}{character vector}

\item{modl_gw_cols}{character vector}
}
\value{
cr.mat matrix This returns the concentration-response matrix
}
\description{
\code{prepCR} preps the concentration-response matrix for a single chemical
}
\details{
ac50,top,w are vectors of length 18 with values for one instance of one chemical

expect top in range [0,100]
ATG = log foldchange top * 25
}

