% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptor_score.R
\name{receptor_score}
\alias{receptor_score}
\title{Receptor Score}
\usage{
receptor_score(x, pathway, aucscale2 = NULL)
}
\arguments{
\item{x}{x}

\item{pathway}{ER or AR}

\item{aucscale2}{Default of NULL will be reset to pathway specific defaults
for scaling AUC values}
}
\value{
score
}
\description{
\code{receptor_score} calculates the receptor score.
}

