% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmat_va.R
\name{tmat_va}
\alias{tmat_va}
\title{tmat_va}
\usage{
tmat_va(pathway, nassay, nreceptor)
}
\arguments{
\item{pathway}{character length 1. Options are "ER" or "AR".}

\item{nassay}{integer length 1}

\item{nreceptor}{integer length 1}
}
\value{
tmat matrix of assay connectivities
}
\description{
\code{tmat_va} returns the connectivity matrix.
}

